feat (agent): support multipe agents running simultaneously
Agents ping Firebase every 10 seconds plus when changing state between offline,
idle, and busy. If an agent fails to ping in 12 seconds, it is assumed to be
offline.

The first agent to successfully execute a transaction against a build tree
updating the build's agent property to its own agent name runs that build.

The web interface displays text and a progress bar indicating how many agents
are currently running a build.

TODO: Implement a system to update a crashed agent's status and set its current
build status to indicate the agent crashed, if needed.

closes #26