feat (agent): build any available branch
Use GitHub for authentication and use the user's credential to access the
GitHub API to get a list of branches for each repo in a build configuration.

The repos listed in a build configuration now have a 'defaultBranchOnly'. If set,
the default branch of that repo of that repo will be cloned for every build
regardless of the selected branch. This is useful for repos that only have a
master branch.

The queue build dialog will show a list of branches that exist in all repos
except those with 'defaultBranchOnly' set to true. The build agent will build
whichever branch is selected.

closes #14, closes #20