fix (agent): use onDisconnect to set agents to offline if crashed
The ping timestamp did not work correctly when the local computer and the agent
had a time difference greater than 12 seconds. This solves that by using
Firebase's native onDisconnect feature to automatically set an agent's status
to offline when it disconnects.

The agent will still sets its status to offline on its own during a clean exit.

fixes #30